Redux - Logic structure
- Re-arrange placement of customer memoized selectors
